Fix addTempRow showing stale name, refactor FileExplorerRow to bring out editing state logic
